Rename Quorum into Strategy
